split Controllers and MediaManager
